+++ realm/realm-library/src/objectServer/java/io/realm/SyncManager.java
@@ -28,6 +28,7 @@ import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -129,6 +130,12 @@ public class SyncManager {
     // Right now it just lives and dies together with the process.
     private static volatile AuthenticationServer authServer = new OkHttpAuthenticationServer();
     private static volatile UserStore userStore;
+
+    // Header configuration
+    private static String globalAuthorizationHeaderName = "Authorization"; // authorization header name if no host-defined header is available
+    private static Map<String, String> hostRestrictedAuthorizationHeaderName = new HashMap<>(); // authorization header name for the given host
+    private static Map<String, String> globalCustomHeaders = new HashMap<>();
+    private static Map<String, Map<String, String>> hostRestrictedCustomHeaders = new HashMap<>();
 
     private static NetworkStateReceiver.ConnectionListener networkListener = new NetworkStateReceiver.ConnectionListener() {
         @Override
@@ -277,6 +284,148 @@ public class SyncManager {
     }
 
     /**
+     * Sets the name of the HTTP header used to send authorization data in when making requests to
+     * all Realm Object Servers used by the app. These servers must have been configured to expect a
+     * custom authorization header.
+     * <p>
+     * The default authorization header is named "Authorization".
+     *
+     * @param headerName name of the header.
+     * @throws IllegalArgumentException if a null or empty header is provided.
+     * @see <a href="https://docs.realm.io/platform/guides/learn-realm-sync-and-integrate-with-a-proxy#adding-a-custom-proxy">Adding a custom proxy</a>
+     */
+    public static synchronized void setAuthorizationHeaderName(String headerName) {
+        checkNotEmpty(headerName, "headerName");
+        authServer.setAuthorizationHeaderName(headerName, null);
+        globalAuthorizationHeaderName = headerName;
+    }
+
+    /**
+     * Sets the name of the HTTP header used to send authorization data in when making requests to
+     * the Realm Object Server running on the defined {@code host}. This server must have been
+     * configured to expect a custom authorization header.
+     * <p>
+     * The default authorization header is named "Authorization".
+     *
+     * @param headerName name of the header.
+     * @param host if this is provided, the authorization header name will only be used on this particular host.
+     *             Example of valid values: "localhost", "127.0.0.1" and "myinstance.us1.cloud.realm.io".
+     * @throws IllegalArgumentException if a {@code null} or empty header and/or host is provided.
+     * @see <a href="https://docs.realm.io/platform/guides/learn-realm-sync-and-integrate-with-a-proxy#adding-a-custom-proxy">Adding a custom proxy</a>
+     */
+
+    public static synchronized void setAuthorizationHeaderName(String headerName, String host) {
+        checkNotEmpty(headerName, "headerName");
+        checkNotEmpty(host, "host");
+        host = host.toLowerCase(Locale.US);
+        authServer.setAuthorizationHeaderName(headerName, host);
+        hostRestrictedAuthorizationHeaderName.put(host, headerName);
+    }
+
+    /**
+     * Adds an extra HTTP header to append to every request to a Realm Object Server.
+     *
+     * @param headerName the name of the header.
+     * @param headerValue the value of header.
+     * @throws IllegalArgumentException if a non-empty {@code headerName} is provided or a null {@code headerValue}.
+     */
+    public static synchronized void addCustomRequestHeader(String headerName, String headerValue) {
+        checkNotEmpty(headerName, "headerName");
+        checkNotNull(headerValue, "headerValue");
+        authServer.addHeader(headerName, headerValue, null);
+        globalCustomHeaders.put(headerName, headerValue);
+    }
+
+    /**
+     * Adds an extra HTTP header to append to every request to a Realm Object Server.
+     *
+     * @param headerName the name of the header.
+     * @param headerValue the value of header.
+     * @param host if this is provided, the this header will only be used on this particular host.
+     *             Example of valid values: "localhost", "127.0.0.1" and "myinstance.us1.cloud.realm.io".
+     * @throws IllegalArgumentException If an non-empty {@code headerName}, {@code headerValue} or {@code host} is provided.
+     */
+    public static synchronized void addCustomRequestHeader(String headerName, String headerValue, String host) {
+        checkNotEmpty(headerName, "headerName");
+        checkNotNull(headerValue, "headerValue");
+        checkNotEmpty(host, "host");
+
+        // Headers
+        host = host.toLowerCase(Locale.US);
+        authServer.addHeader(headerName, headerValue, host);
+        Map<String, String> headers = hostRestrictedCustomHeaders.get(host);
+        if (headers == null) {
+            headers = new LinkedHashMap<>();
+            hostRestrictedCustomHeaders.put(host, headers);
+        }
+        headers.put(headerName, headerValue);
+    }
+
+    /**
+     * Adds extra HTTP headers to append to every request to a Realm Object Server.
+     *
+     * @param headers map of (headerName, headerValue) pairs.
+     * @throws IllegalArgumentException If any of the headers provided are illegal.
+     */
+    public static synchronized void addCustomRequestHeaders(@Nullable Map<String, String> headers) {
+        if (headers != null) {
+            for (Map.Entry<String, String> entry : headers.entrySet()) {
+                addCustomRequestHeader(entry.getKey(), entry.getValue());
+            }
+        }
+    }
+
+    /**
+     * Adds extra HTTP headers to append to every request to a Realm Object Server.
+     *
+     * @param headers map of (headerName, headerValue) pairs.
+     * @param host if this is provided, the this header will only be used on this particular host.
+     *             Example of valid values: "localhost", "127.0.0.1" and "myinstance.us1.cloud.realm.io".
+     * @throws IllegalArgumentException If any of the headers provided are illegal.
+     */
+    public static synchronized void addCustomRequestHeaders(@Nullable Map<String, String> headers, String host) {
+        if (Util.isEmptyString(host)) {
+            throw new IllegalArgumentException("Non-empty 'host' required");
+        }
+        host = host.toLowerCase(Locale.US);
+        if (headers != null) {
+            for (Map.Entry<String, String> entry : headers.entrySet()) {
+                addCustomRequestHeader(entry.getKey(), entry.getValue(), host);
+            }
+        }
+    }
+
+    /**
+     * Returns the authentication header name used for the http request to the given url.
+     *
+     * @param objectServerUrl Url to get header for.
+     * @return the authorization header name used by http requests to this url.
+     */
+    public static synchronized String getAuthorizationHeaderName(URI objectServerUrl) {
+        String host = objectServerUrl.getHost().toLowerCase(Locale.US);
+        String hostRestrictedHeader = hostRestrictedAuthorizationHeaderName.get(host);
+        return (hostRestrictedHeader != null) ? hostRestrictedHeader : globalAuthorizationHeaderName;
+    }
+
+    /**
+     * Returns all the custom headers added to requests to the given url.
+     *
+     * @return all defined custom headers used when making http requests to the given url.
+     * f
+     */
+    public static synchronized Map<String, String> getCustomRequestHeaders(URI serverSyncUrl) {
+        Map<String, String> headers = new LinkedHashMap<>(globalCustomHeaders);
+        String host = serverSyncUrl.getHost().toLowerCase(Locale.US);
+        Map<String, String> hostHeaders = hostRestrictedCustomHeaders.get(host);
+        if (hostHeaders != null) {
+            for (Map.Entry<String, String> entry : hostHeaders.entrySet()) {
+                headers.put(entry.getKey(), entry.getValue());
+            }
+        }
+        return headers;
+    }
+
+    /**
      * Remove the wrapped Java session.
      * @param syncConfiguration configuration object for the synchronized Realm.
      */
@@ -303,6 +452,7 @@ public class SyncManager {
      * @return the all valid sessions belonging to the user.
      */
     static List<SyncSession> getAllSessions(SyncUser syncUser) {
+        //noinspection ConstantConditions
         if (syncUser == null) {
             throw new IllegalArgumentException("A non-empty 'syncUser' is required.");
         }
@@ -553,6 +703,18 @@ public class SyncManager {
         }
     }
 
+    private static void checkNotEmpty(String headerName, String varName) {
+        if (Util.isEmptyString(headerName)) {
+            throw new IllegalArgumentException("Non-empty '" + varName +"' required.");
+        }
+    }
+
+    private static void checkNotNull(@Nullable String val, String varName) {
+        if (val == null) {
+            throw new IllegalArgumentException("Non-null'" + varName +"' required.");
+        }
+    }
+
     /**
      * Resets the SyncManger and clear all existing users.
      * This will also terminate all sessions.
@@ -562,6 +724,11 @@ public class SyncManager {
     static synchronized void reset() {
         nativeReset();
         sessions.clear();
+        hostRestrictedAuthorizationHeaderName.clear();
+        globalAuthorizationHeaderName = "Authorization";
+        hostRestrictedCustomHeaders.clear();
+        globalCustomHeaders.clear();
+        authServer.clearCustomHeaderSettings();
     }
 
     /**